SplitSliceIntoMaxLengthSlices now takes max length as a uint

+++ util/slice.go
@@ -3,8 +3,8 @@ package util
 import "math"
 
 // Slices a slice into slices that are all of the maximum length (or less)
-func SplitSliceIntoMaxLengthSlices[T any](original []T, maxLength int) [][]T {
-	originalLen := len(original)
+func SplitSliceIntoMaxLengthSlices[T any](original []T, maxLength uint) [][]T {
+	originalLen := uint(len(original))
 
 	resultsCountFloat := math.Ceil(float64(originalLen) / float64(maxLength))
 	resultsCount := int(resultsCountFloat)
@@ -12,7 +12,7 @@ func SplitSliceIntoMaxLengthSlices[T any](original []T, maxLength int) [][]T {
 	results := make([][]T, 0, resultsCount)
 
 	for i := 0; i < resultsCount; i++ {
-		startIdx := maxLength * i
+		startIdx := maxLength * uint(i)
 		endIdx := startIdx + maxLength
 
 		if endIdx > originalLen {
